% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSurrogates.R
\name{addSurrogates}
\alias{addSurrogates}
\title{Add surrogate information that was created by getTreeranger}
\usage{
addSurrogates(RF, trees, s, Xdata, num.threads)
}
\arguments{
\item{RF}{random forest object created by ranger (with keep.inbag=TRUE).}

\item{trees}{list of trees created by getTreeranger.}

\item{s}{Predefined number of surrogate splits (it may happen that the actual number of surrogate splits differes in individual nodes). Default is 1 \% of no. of variables.}

\item{Xdata}{data without the dependent variable.}

\item{num.threads}{number of threads used for parallel execution. Default is number of CPUs available.}
}
\value{
a list with trees containing of lists of nodes with the elements:
\itemize{
\item nodeID: ID of the respective node (important for left and right daughters in the next columns)
\item leftdaughter: ID of the left daughter of this node
\item rightdaughter: ID of the right daughter of this node
\item splitvariable: ID of the split variable
\item splitpoint: splitpoint of the split variable
\item status: "0" for terminal and "1" for non-terminal
\item layer: layer information (0 means root node, 1 means 1 layer below root, etc)
\item surrogate_i: numbered surrogate variables (number depending on s)
\item adj_i: adjusted agreement of variable i
}
}
\description{
This function adds surrogate variables and adjusted agreement values to a forest that was created by getTreeranger.
}
