% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_clusters.R
\name{build.clusters}
\alias{build.clusters}
\title{Apply cluster analysis to build variable groups}
\usage{
build.clusters(rel, hcmethod = "ward.D")
}
\arguments{
\item{rel}{a list containing variables, surr.res, threshold, and var. This is the output of \link[SurrogateMinimalDepth]{var.relations} function.}

\item{hcmethod}{the hierarchical clustering method that is used. (see \link[linkcomm]{getLinkCommunities})}
}
\value{
a data frame containing the variable names and their associated clusters.
}
\description{
This function generates variables groups of relation information that was obtained by \link[SurrogateMinimalDepth]{var.relations} function applying
\link[linkcomm]{getLinkCommunities}.
}
\examples{
# read data
data("SMD_example_data")
x = SMD_example_data[,2:ncol(SMD_example_data)]
y = SMD_example_data[,1]

\donttest{
 # apply SMD to obtain random forest with surrogate variables (usually more trees are needed)
 set.seed(42)
 res.smd = var.select.smd(x = x, y = y, ntree = 10)

 # investigate variable relations
 variables = res.smd$forest$allvariables
 rel=var.relations(forest = res.smd$forest, variables = variables, candidates = variables, t = 10)
 # built groups based on mean adjusted agreement
 groups = build.clusters(rel)
}

}
