% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_mgi.R
\name{calculate.mgi.perm}
\alias{calculate.mgi.perm}
\title{This is an internal function}
\usage{
calculate.mgi.perm(r = 1, adj.agree_perm, air, allvariables)
}
\description{
This is an internal function
}
\keyword{internal}
