% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_mir.R
\name{calculate.mir.perm}
\alias{calculate.mir.perm}
\title{calculate.mir.perm}
\usage{
calculate.mir.perm(r = 1, adj.agree.perm, air, variables)
}
\description{
This is an internal function
}
\keyword{internal}
