% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.surrogates.R
\name{count.surrogates}
\alias{count.surrogates}
\title{Count surrogate variables}
\usage{
count.surrogates(trees)
}
\arguments{
\item{trees}{list of trees that was generated by getTreeranger function and layers, surrogate variables, and adjusted agreement values were added by addLayer and getSurrogates functions}
}
\value{
List with the following components:
\itemize{
\item s.a: total average number of surrogate variables
\item s.l: average number of surrogate variables in the respective layers
}
}
\description{
This function counts surrogate variables and returns the total average number of surrogate variables and the average number of surrogate variables of the respective layers.
This is necessary since the actual number of surrogate splits can be lower than the predefined number (when less surrogate splits outperform the majority rule).
}
