% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeranger.R
\name{getTreeranger}
\alias{getTreeranger}
\title{Get a list of structured trees for ranger}
\usage{
getTreeranger(RF, num.trees)
}
\arguments{
\item{RF}{random forest object created by ranger (with keep.inbag=TRUE)}

\item{num.trees}{number of trees}
}
\value{
a list with trees. Each row of the list elements corresponds to a node of the respective tree and the columns correspond to:
\itemize{
\item nodeID: ID of the respective node (important for left and right daughters in the next columns)
\item leftdaughter: ID of the left daughter of this node
\item rightdaughter: ID of the right daughter of this node
\item splitvariable: ID of the split variable
\item splitpoint: splitpoint of the split variable (for categorical variables this is a comma separated lists of values, representing the factor levels (in the original order) going to the right)
\item status: "0" for terminal and "1" for non-terminal
}
}
\description{
This functions creates a list of trees for ranger objects similar as getTree function does for random Forest objects.
}
