% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanAdjAgree.R
\name{meanAdjAgree}
\alias{meanAdjAgree}
\title{Calculate mean adjusted agreement to investigate variables relations}
\usage{
meanAdjAgree(
  trees,
  variables,
  allvariables,
  candidates,
  t,
  s.a,
  select.var,
  num.threads = NULL
)
}
\arguments{
\item{trees}{list of trees created by getTreeranger, addLayer and addSurrogate.}

\item{variables}{vector of variable names.}

\item{allvariables}{vector of all variable names (strings)}

\item{candidates}{vector of variable names (strings) that are candidates to be related to the variables (has to be contained in allvariables)}

\item{t}{variable to calculate threshold. Default is 3.}

\item{s.a}{average number of surrogate variables (ideally calculated by count.surrogates function).}

\item{select.var}{set False if only relations should be calculated and no related variables should be selected.}

\item{num.threads}{number of threads used for parallel execution. Default is number of CPUs available.}
}
\value{
a list containing:
\itemize{
\item variables: the variables to which relations are investigated
\item surr.res: matrix with mean adjusted agreement values and variables investigated in rows and candidate variables in columns
\item threshold: the threshold used to create surr.var from surr.res
\item surr.var: binary matrix showing if the variables are related (1) or non-related (0) with variables in rows and candidates in columns.
}
}
\description{
This is the main function of var.relations function.
}
