% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindep.R
\name{mindep}
\alias{mindep}
\title{Execute minimal depth variable importance}
\usage{
mindep(variables, trees)
}
\arguments{
\item{variables}{vector of variable names}

\item{trees}{list of trees that was generated by getTreeranger and layers functions}
}
\value{
List with the following components:
\itemize{
\item depth: mean minimal depth for each variable
\item selected: variables has been selected (1) or not (0),
\item threshold: the threshold that is used for the selection
}
}
\description{
This function determines the minimal depth of variables from a forest that is created by getTreeranger, and addLayer functions.
}
