% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_mir.R
\name{p.relation}
\alias{p.relation}
\title{calculate.mir.perm}
\usage{
p.relation(l = 1, adj.agree, adj.agree.perm, adj.agree.corr, variables)
}
\description{
This is an internal function
}
\keyword{internal}
