% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrmindep.R
\name{surrmindep}
\alias{surrmindep}
\title{Execute surrogate minimal depth variable importance}
\usage{
surrmindep(forest, s.l)
}
\arguments{
\item{forest}{a list containing allvariables and trees. Allvariables is a vector of all variable names in the original data set (strings). Trees is a list of trees that was generated by getTreeranger, addLayer, and getSurrogates functions.}

\item{s.l}{Number of average surrogate variables in the respective layers. (use count.surrogate function to get it)}
}
\value{
List with the following components:
\itemize{
\item depth: mean surrogate minimal depth for each variable
\item selected: variables has been selected (1) or not (0),
\item threshold: the threshold that is used for the selection
}
}
\description{
This function determines the surrogate minimal depth of variables from a forest that is created by getTreeranger, addLayer and getSurrogates functions.
}
