% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_md.R
\name{var.select.md}
\alias{var.select.md}
\title{Variable selection with Minimal Depth (MD)}
\usage{
var.select.md(x, y, ntree = 500, type = "regression", mtry = NULL,
  min.node.size = 1, num.threads = NULL)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed)}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used). For survival forests this is the time variable.}

\item{ntree}{Number of trees. Default is 500.}

\item{type}{Mode of prediction ("regression" or "classification"). Default is regression.}

\item{mtry}{Number of variables to possibly split at in each node. Default is no. of variables^(3/4) as recommended by Ishwaran.}

\item{min.node.size}{Minimal node size. Default is 1.}

\item{num.threads}{number of threads used for parallel execution. Default is number of CPUs available.}

\item{status}{status variable, only applicable to survival data. Use 1 for event and 0 for censoring.}
}
\value{
List with the following components:
\itemize{
\item info: list with results from mindep function:
\itemize{
\item depth: mean minimal depth for each variable
\item selected: variables has been selected (1) or not (0),
\item threshold: the threshold that is used for the selection
}
\item var: vector of selected variables

\item trees: list of trees that was created by getTreeranger, and addLayer functions and that was used for minimal depth variable importance
}
}
\description{
This function executes MD applying \link[ranger]{ranger} for random forests generation and is a reimplementation of \link[randomForestSRC]{var.select} from randomForestSRC package.
}
\examples{
# read data
data("SMD_example_data")

\donttest{
# select variables (usually more trees are needed)
res = var.select.md(x=SMD_example_data[,2:ncol(SMD_example_data)],y=SMD_example_data[,1],ntree=10)
res$var
}

}
\references{
\itemize{
  \item Ishwaran, H. et al. (2011) Random survival forests for high-dimensional data. Stat Anal Data Min, 4, 115–132. \url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sam.10103}
  \item Ishwaran, H. et al. (2010) High-Dimensional Variable Selection for Survival Data. J. Am. Stat. Assoc., 105, 205–217. \url{http://www.ccs.miami.edu/~hishwaran/papers/IKGML.JASA.2010.pdf}
  }
}
