% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_clusters.R
\name{build.clusters}
\alias{build.clusters}
\title{Apply cluster analysis to build variable groups}
\usage{
build.clusters(rel, hcmethod = "ward.D")
}
\arguments{
\item{rel}{a list containing variables, surr.res, threshold, and var. This is the output of \link[SurrogateMinimalDepth]{var.relations} function.}

\item{hcmethod}{the hierarchical clustering method that is used. (see \link[linkcomm]{getLinkCommunities})}
}
\value{
a data frame containing the variable names and their associated clusters.
}
\description{
This function generates variables groups of relation information that was obtained by \link[SurrogateMinimalDepth]{var.relations} function applying
\link[linkcomm]{getLinkCommunities}.
}
\examples{
# read data
data("SMD_example_data")

\donttest{
 # get trees and variable names
 x = SMD_example_data[,2:ncol(SMD_example_data)]
 y = SMD_example_data[,1]
 allvariables = colnames(x)# extract variables names
 nvar = length(allvariables)   # count number of variables
 set.seed(42)
 RF = ranger::ranger(data = SMD_example_data, dependent.variable.name = "y", num.trees = 10, keep.inbag = TRUE,mtry = floor(nvar^(3/4)), min.node.size = 1)
 trees = getTreeranger(RF = RF, num.trees = 10)
 trees.lay = addLayer(trees)
 trees.surr = addSurrogates(RF = RF, trees = trees.lay, s = 10, Xdata = x, num.threads = NULL)

 # investigate variable relations
 rel=var.relations(forest = list(trees = trees.surr, allvariables = allvariables), variables = allvariables, candidates = allvariables, t = 10)
 groups = build.clusters(rel)
}

}
